userspace-dp: extract afxdp::binding_state out of afxdp::umem (#6436)
umem/mod.rs was nominally the UMEM memory-region module but ~85% of
it was binding runtime state: the 153-field BindingLiveState atomics
cluster, the PendingTxAdmission RAII single-release token, the hot
per-packet TX enqueue paths, and the HA session-delta loss-latch
buffer, plus the snapshot/debug_state/profile renderers that exist
solely to serve BindingLiveState. A module that mixes memory mapping
with ring-state counters silently degrades reviewability and
inlining, so split it before the next change lands on it
(docs/engineering-style.md modularity discipline).

Pure code-motion within the crate:

- binding_state/mod.rs holds the BindingLiveState struct (field
  order verbatim — drop order is load-bearing where documented),
  its new() + setter/counter impl block, SharedUmemLiveStatus,
  FlowWorkerMapSnapshot, and the pub(in crate::afxdp) re-exports.
- binding_state/tx_inbox.rs holds the bounded lock-free MPSC
  redirect inbox and its linearizable admission counter: enqueue_tx,
  enqueue_tx_owned, try_enqueue_tx_owned, take_pending_tx_into,
  try_reserve_mirror_tx_owned, the PendingTxAdmission RAII token,
  and PENDING_TX_INBOX_HARD_CAP.
- binding_state/latency.rs holds bucket_index_for_ns, the
  DRAIN_HIST_BUCKETS / TX_SUBMIT_LAT_BUCKETS wire-contract constants
  with their const asserts, the producer-local REDIRECT_SAMPLE_SEQ
  TLS sampler, and TX_SIDECAR_UNSTAMPED.
- binding_state/session_delta.rs holds the RPC-fallback
  pending_session_deltas buffer and the #5290 delta_loss_pending
  loss-of-sync latch.
- snapshot.rs / debug_state.rs / profile.rs move verbatim (git
  rename; only stale umem.rs path citations in comments updated).
- The #4667 per-concern umem/tests/ layout maps 1:1 onto
  binding_state/tests/; umem/tests/ keeps the mmap_area concern.

umem/ shrinks to ~120 lines of pure memory region (MmapArea,
WorkerUmem{,Inner,Pool}).

Visibility is preserved exactly: pub(super) from binding_state
resolves to pub(in crate::afxdp) just as it did from umem; methods
moving into submodules are re-declared pub(in crate::afxdp), the
same effective visibility they had. REDIRECT_SAMPLE_SEQ stays
module-private; next_redirect_sample is pub(super) within
binding_state. The REDIRECT_SAMPLE_MASK re-export is cfg(test)-gated
(only the latency_buckets tests consume it), matching the existing
WorkerTimers gating precedent — warning parity 159/159 (cargo check)
and 120/120 (check --tests) vs the 023f17a60 base.

asm-diff on the canonical target/release/xpf-userspace-dp binary
(objdump -Cd, baseline vs PR, address/symbol-name normalized):
push_redirect_inbox 89/89 instructions identical,
take_pending_tx_into 135/135 identical,
drop_in_place<PendingTxAdmission> 10/10 identical (RAII
disarm-on-push + AcqRel release byte-identical), enqueue_tx_owned
254/254 with the sole delta being the REDIRECT_SAMPLE_SEQ TLS slot
offset (module-rename TLS relocation, same instruction).

Validation: cargo check (release + debug, with --tests) clean at
warning parity; cargo test --release --bins --tests --
--test-threads=1 (the make test-rust command) green twice, 4250
tests, 0 failures. No cluster smoke: pure code-motion refactor with
no behavior change, so there is no feature-level lane to run.

Closes #6436
